netlify error states ERROR in ./src/pages/Home.js, so moved Home back to components in hopes it helps build
